+++ pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "TP_release_converter"
+version = "0.0.1"
+authors = [
+  { name="Angileri", email="lorenzo.angileri@isen.yncrea.fr" },
+]
+description = "apprentissage de relaese de package"
+readme = "README.md"
+requires-python = ">=3.7"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
